refactor migration folders
