+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  "config": path.resolve('./src/database', 'config.js'),
+  "config": path.resolve('./src/database/config/config.js'),
   "models-path": path.resolve('./src/models'),
-  "seeders-path": path.resolve('./src/seeders'),
-  "migrations-path": path.resolve('./src/migrations')
+  "seeders-path": path.resolve('./src/database/seeders'),
+  "migrations-path": path.resolve('./src/database/migrations')
 };
