+++ src/mock/mockedStrategy.js
