+++ src/models/index.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import Sequelize from 'sequelize';
-import configs from '../database/config';
+import configs from '../database/config/config';
 
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
