+++ src/tests/testData/sampleData.js
@@ -2,16 +2,33 @@ export const multiRequest = {
   source: 'Nigeria',
   destination: 'Abuja, USA, POLAND',
   tripType: 'multi-city',
-  travelDate: '2038-01-19 03:14:07 ',
-  returnDate: '2038-01-19 03:14:07',
+  travelDate: Date.now(),
+  returnDate: Date.now(),
   reason: 'reason',
   accommodation: 'accommodation'
 };
 
 export const missingRequiredField = {
   tripType: 'multi-city',
-  travelDate: '2038-01-19 03:14:07 ',
-  returnDate: '2038-01-19 03:14:07',
+  travelDate: Date.now(),
+  returnDate: Date.now(),
+  reason: 'reason',
+  accommodation: 'accommodation'
+};
+
+export const user = {
+  firstName: 'Samuel',
+  lastName: 'koroh',
+  email: 'user1@gmail.com',
+  password: 'Ice5m5am0a843r03'
+};
+
+export const oneWayTrip = {
+  source: 'Lagos',
+  tripType: 'one-way',
+  destination: 'Abuja',
+  travelDate: Date.now(),
+  returnDate: Date.now(),
   reason: 'reason',
   accommodation: 'accommodation'
 };
